--- a/examples/olmoe.py
+++ b/examples/olmoe.py
@@ -43,12 +43,7 @@ def fetch_weights() -> dict[str, Tensor]:
   m3 = Tensor.from_url("https://huggingface.co/allenai/OLMoE-1B-7B-0924/resolve/main/model-00003-of-00003.safetensors").to(Device.DEFAULT)
   return {**nn.state.safe_load(m1), **nn.state.safe_load(m2), **nn.state.safe_load(m3)}
 
-def filter_layers(state, layers):
-  lay = [f"model.layers.{i}." for i in range(layers)]
-  return {k: v for k, v in state.items() if not k.startswith("model.layers.") or any(k.startswith(l) for l in lay)}
-
 if __name__ == "__main__":
-  LAYERS = 8
   if getenv("TORCH"):
     from transformers import OlmoeForCausalLM, AutoTokenizer
     model = OlmoeForCausalLM.from_pretrained("allenai/OLMoE-1B-7B-0924")
@@ -60,15 +55,13 @@ if __name__ == "__main__":
     exit(0)
 
   with Timing("create model: "):
-    # model = Transformer(n_layers=16, dim=2048, hidden_dim=1024, n_heads=16, norm_eps=1e-5, qk_norm=1e-5, max_context=1024,
-    #                     vocab_size=50304, feed_forward=functools.partial(MixtureFeedForward, 64, 8))
-    model = Transformer(n_layers=LAYERS, dim=2048, hidden_dim=1024, n_heads=16, norm_eps=1e-5, qk_norm=1e-5, max_context=1024,
+    model = Transformer(n_layers=16, dim=2048, hidden_dim=1024, n_heads=16, norm_eps=1e-5, qk_norm=1e-5, max_context=1024,
                         vocab_size=50304, feed_forward=functools.partial(MixtureFeedForward, 64, 8))
     model_state_dict = nn.state.get_state_dict(model)
     del model_state_dict['freqs_cis']
 
   with Timing("load weights to GPU: "):
-    nhf_state = convert_from_huggingface(filter_layers(fetch_weights(), LAYERS), model, 16, 16)
+    nhf_state = convert_from_huggingface(fetch_weights(), model, 16, 16)
     # NOTE: i'm not sure this actually needs float32, it may just change the type of things downstream from it. but doesn't match torch w/o this
     for needs_float32 in ['tok_embeddings.weight']: nhf_state[needs_float32] = nhf_state[needs_float32].float()
   print(f"ram used: {GlobalCounters.mem_used/1e9:.2f} GB")
@@ -98,22 +91,6 @@ if __name__ == "__main__":
     start_pos += 1
     print(toks)
     print(tokenizer.decode(toks))
-
-  ### LAYER = 2
-  # old: JITBEAM=2 fastest token 34.77 ms, 28.8 tok/s
-  # new: JITBEAM=2 fastest token 9.49 ms, 105.4 tok/s  3.7x speedup
-
-  ### LAYER = 4
-  # old: JITBEAM=2 fastest token 64.87 ms, 15.4 tok/s
-  # new: JITBEAM=2 fastest token 14.19 ms, 70.5 tok/s  4.7x speedup
-
-  ### LAYER = 8
-  # old: JITBEAM=2 fastest token 123.58 ms, 8.1 tok/s
-  # new: JITBEAM=2 fastest token 23.02 ms, 43.4 tok/s  5.4x speedup
-
-  ### LAYER = 16
-  # ??? 6x speedup?
-
   print(f"fastest token {min(timings)*1e3:.2f} ms, {1/min(timings):.1f} tok/s")
 
   if temperature == 0: