```
  ### LAYER = 2
  # old: JITBEAM=2 fastest token 34.77 ms, 28.8 tok/s
  # new: JITBEAM=2 fastest token 9.49 ms, 105.4 tok/s  3.7x speedup

  ### LAYER = 4
  # old: JITBEAM=2 fastest token 64.87 ms, 15.4 tok/s
  # new: JITBEAM=2 fastest token 14.19 ms, 70.5 tok/s  4.7x speedup

  ### LAYER = 8
  # old: JITBEAM=2 fastest token 123.58 ms, 8.1 tok/s
  # new: JITBEAM=2 fastest token 23.02 ms, 43.4 tok/s  5.4x speedup

  ### LAYER = 16
  # ??? 6x speedup?
```

pretty beefy speed up with magical fuse arange
saw `beautiful_cifar.py` do it like this and I just copied it over.
Not sure if this fixes the "multiplying by 0 instead of not computing" gotta investigate.