@chenyuxyz 2025-03-30
for the whole thing (16 layers) i got 16.8 tok/s -> 64 tok/s