@geohot 2025-03-31
I'm getting 53.3 tok/s on M3 Max with 400 GB/s

The model has 1.3B active params, for 2.6GB of weights, theoretical max is 400/2.6 = 153.8 tok/s.

So this is close, can you get 76.9 tok/s for the bounty?