@geohot 2025-05-05
`JITBEAM=2 python3 examples/olmoe.py`

`fastest token 17.82 ms, 56.1 tok/s`

Do I need other flags?