@geohotstan 2025-05-06
Huh, that's weird.
`JITBEAM=2 python3 examples/olmoe.py` should've ran it with `FUSE_ARANGE=1`
I think `57.5 tok/s` is just the new speed since there was no speed regression (fuse arange was definitely applied)

Let me investigate.