@chenyuxyz 2025-08-17
unlock the bounty for this one so that more people can try, i think previous benefits are covered by FUSE_ARANGE which is the default now